% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.ppgram}
\alias{as.data.frame.ppgram}
\title{Convert pppgram to Data Frame}
\usage{
\method{as.data.frame}{ppgram}(x, ...)
}
\arguments{
\item{x}{pppgram-object}

\item{...}{passed on to `data.frame`}
}
\description{
Convert pppgram to Data Frame
}
