% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.pppgram}
\alias{as.data.frame.pppgram}
\title{#' Wrapper for Rotation Averaging pppgram
#' 
#' @param x pppgram-object
#' @param ... passed on to spatstat::rotmean
#' 
#' @export
rotmean.pppgram <- function(x, ...) {
  y <- as.im(x)
  rotmean(y, ...)
}
Convert pppgram to Data Frame}
\usage{
\method{as.data.frame}{pppgram}(x, ...)
}
\arguments{
\item{x}{pppgram-object}

\item{...}{passed on to `data.frame`}
}
\description{
#' Wrapper for Rotation Averaging pppgram
#' 
#' @param x pppgram-object
#' @param ... passed on to spatstat::rotmean
#' 
#' @export
rotmean.pppgram <- function(x, ...) {
  y <- as.im(x)
  rotmean(y, ...)
}
Convert pppgram to Data Frame
}
