% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.im.pppgram}
\alias{as.im.pppgram}
\title{Convert pppgram to im-object}
\usage{
\method{as.im}{pppgram}(x, ...)
}
\arguments{
\item{x}{pppgram-object}

\item{...}{ignored}
}
\description{
Convert pppgram to im-object
}
