% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pp.R
\name{check_pp}
\alias{check_pp}
\title{Check Input Pattern Format}
\usage{
check_pp(x)
}
\arguments{
\item{x}{pattern candidate}
}
\description{
Check Input Pattern Format
}
\details{
Only support spatstat::ppp objects with rectangular windows.
}
