% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoretical-sdf.R
\name{maternII_generator_intensity}
\alias{maternII_generator_intensity}
\title{Solve for the MaternII generator intensity}
\usage{
maternII_generator_intensity(target_intensity, hc)
}
\arguments{
\item{target_intensity}{Desired intensity in the end}

\item{hc}{the hard core parameter}
}
\description{
Solve for the MaternII generator intensity
}
