% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoretical-sdf.R
\name{pcf_maternII}
\alias{pcf_maternII}
\title{Matern type II pcf}
\usage{
pcf_maternII(r = seq(0, 2, l = 50), target_intensity = 50, hc = 0.05)
}
\arguments{
\item{r}{spatial scales}

\item{target_intensity}{intensity of the process (not the generator)}

\item{hc}{hard-core distance}
}
\description{
Isotropic pair correlation function of Matern type II
}
\details{
Generator, or dominant Poisson to be thinned, intensity will be solved 
and an error will be given if target intensity is too high given hc distance.
}
