% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoretical-sdf.R
\name{pcf_to_sdf}
\alias{pcf_to_sdf}
\title{Generic function to integrate isotropic pcf to sdf}
\usage{
pcf_to_sdf(g, r, w, lam)
}
\arguments{
\item{g}{pcf values at r}

\item{r}{spatial arguments}

\item{w}{wavenumber arguments}

\item{lam}{intensity}
}
\description{
Numerical Hankel transform of pcf to sdf
}
\details{
Not the most robust Hankel transform, so use with care.
}
