% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Bartlett's Periodogram for 2D Point Patterns}
\usage{
periodogram(x, omega = NULL, ..., M = 25, debias_early = !is.null(omega))
}
\arguments{
\item{x}{ppp-object, i.e. spatstat's point pattern, with rectangular window.}

\item{omega}{The wavenumber grid}

\item{...}{ignored}

\item{M}{max steps in Fourier wavenumber grid when omega missing.}

\item{debias_early}{Do debiasing before squaring (default: TRUE if omega given)? No difference for square windows and Fourier
grids, but great reduction in bias for non-square windows (future) or non-Fourier grids.}
}
\description{
Bartlett's Periodogram for 2D Point Patterns
}
\details{
We estimate the spectral density defined by Bartlett 1964 for a stationary process with 
intensity $\lambda$ and a pair correlation $g$ as

\deqn{\mathcal{F}(\omega) = \lambda + \lambda^2\int_{R^d}[g(z)-1]e^{-2\pi i\omega^T z}dz}

using a "periodogram", viz. a quadratic transformation of the Fourier transform of the data. 

Note how we parameterise the Fourier transform using wavenumbers.

No scaling of the pattern is conducted, only centering to 0 for computational simplicity
(does not affect the estimate as pattern assumed stationary).

The wavenumbers that the spectrum is estimated are given by `omega`:

1) If omega is a vector, we expand it to d-dimensional grid using expand.grid. 

2) If omega is a column matrix of dimension m x d, each row vector is
interpreted as a wavenumber.

If omega is missing, we will expand -M:M/l where l is the minimum sidelength of the data window.

If `debias_early = TRUE`, the systematic bias component due to non-centered stochastic 
process is cancelled before squaring the discrete Fourier transform.
}
\seealso{
\code{\link[=periodogram_mt]{periodogram_mt}} for multi-tapered peridograms; 
\code{\link{smoothen.periodogram}} for operations on the estimated periodograms.
}
