% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pppgram}
\alias{plot.pppgram}
\title{Plot 2D Periodogram Estimate}
\usage{
\method{plot}{pppgram}(x, ...)
}
\arguments{
\item{x}{pppgram-object}

\item{...}{passed on to plot.im of spatstat}
}
\description{
Plot 2D Periodogram Estimate
}
