% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram_iso.R
\name{ppgram_iso_bias}
\alias{ppgram_iso_bias}
\title{Isotropic Periodogram Bias}
\usage{
ppgram_iso_bias(sl, t, taper_a, n = c(150, 200))
}
\arguments{
\item{sl}{window sidelenghts}

\item{t}{wavenumber magnitudes}

\item{taper_a}{parameter of the squared-exponential taper. if 0, no taper.}

\item{n}{grid lengths for numerical integration, order (ang, r)}
}
\description{
Assuming a rectangular window, 2D
}
