% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring-pass-filter.R
\name{ringpass}
\alias{ringpass}
\title{Ring-pass Filter}
\usage{
ringpass(x, a = 0.1, b = 0.2, nx = 2^4, minus = FALSE, fft = TRUE)
}
\arguments{
\item{x}{point pattern, 2D, such as `ppp`}

\item{a}{lower annulus radius}

\item{b}{upper annulus radius}

\item{nx}{resolution, in x-dimension}

\item{minus}{development, will drop points at edge if fft = FALSE}

\item{fft}{use FFT-based approximation? default: TRUE}
}
\value{
spatstat's `im` object
}
\description{
Ring-pass Filter
}
\details{
y-dimension resolution computed to have square pixels

The non-fft version is more accurate, but slow so if `fft=FALSE` keep `nx` low.

`fft=TRUE` will buffer the window and then clip back to avoid circularity issues (toroidality).
}
