% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoretical-sdf.R
\name{sdf_maternII}
\alias{sdf_maternII}
\title{Spectral density function of Matern type II process}
\usage{
sdf_maternII(k, lambda, hc, nint = 2000)
}
\arguments{
\item{k}{wavenumbers}

\item{lambda}{Target intensity}

\item{hc}{hard core range}

\item{nint}{numerical integration scheme steps}
}
\value{
spectral density function scaled by intensity
}
\description{
Spectral density function of Matern type II process
}
\details{
The parametrisation takes the desired intensity `lambda`, not the generator intensity.
The generator intensity will be solved from `lambda` and `hc`, 
and will fail if the combination is not possible.
}
