% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoretical-sdf.R
\name{sdf_thomas}
\alias{sdf_thomas}
\title{Spectral density function of Thomas process}
\usage{
sdf_thomas(k, lambda, mu, sigma)
}
\arguments{
\item{k}{wavenumbers}

\item{lambda}{Target process intensity (not the generator intensity)}

\item{mu}{Offspring intensity}

\item{sigma}{Cluster stdev}
}
\value{
Isotropic spectral density function scaled by intensity
}
\description{
Isotropic sdf of the (modified) Thomas point process
}
\details{
The parametrisation takes the target intensity, not the generator intensity (usually `kappa`). 
But it does not really matter here, scaling by intensity removes connection to 
lambda or kappa, only mu remains.
}
