% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Smooth.pppgram}
\alias{Smooth.pppgram}
\title{Wrapper for Smoothing Periodogram}
\usage{
\method{Smooth}{pppgram}(x, ...)
}
\arguments{
\item{x}{pppgram-object}

\item{...}{passed on to spatstat::blur}
}
\description{
Wrapper for Smoothing Periodogram
}
\details{
this is convenience function for doing blur(as.im(x), ...) so that the result is restored as pppgram-object.
}
