% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram_iso.R
\name{periodogram_iso}
\alias{periodogram_iso}
\title{Isotropic Periodogram Estimator}
\usage{
periodogram_iso(x, t, ..., normalise = TRUE)
}
\arguments{
\item{x}{point pattern ppp-object with a rectangular window}

\item{t}{The wavenumber magnitude grid ("frequencies")}

\item{...}{ignored}

\item{normalise}{Divide by lambda? Poisson process will then be approximately constant 1.}

\item{debias}{Apply debiasing after estimation? Will reduce bias near 0, but might lead to a negative estimates.}
}
\description{
Isotropic Periodogram Estimator
}
\details{
Basically, the Hankel-transform of the data.
}
